moved concurrency stamp hooks to Sienar.Data namespace
