+++ src/Sienar.Architecture.EntityFramework/Extensions/SienarEntityFrameworkServiceCollectionExtensions.cs
@@ -55,7 +55,7 @@ public static class SienarEntityFrameworkServiceCollectionExtensions
 		where TRepository : class, IRepository<TEntity>
 		where TRepositoryImplementation : class, TRepository
 	{
-		self.TryAddScoped<IBeforeAction<TEntity>, ConcurrencyStampUpdateHook<TEntity>>();
+		self.TryAddScoped<IBeforeAction<TEntity>, ConcurrencyStampUpdater<TEntity>>();
 		self.TryAddScoped<IStateValidator<TEntity>, ConcurrencyStampValidator<TEntity>>();
 		self.TryAddScoped<IEntityFrameworkFilterProcessor<TEntity>, TFilterProcessor>();
 		self.TryAddScoped<TRepository, TRepositoryImplementation>();
